Create Symmetric Tree(Easy)
Given a root of a Binary Tree. Your task is to check whether it is Symmetric or not, i.e. whether the binary tree is a Mirror image of itself.

Examples:

Input: root[] = [1, 2, 2, 3, 4, 4, 3]
   ex-1_1
Output: True
Explanation: Tree is mirror image of itself i.e. tree is symmetric.
Input: root[] = [1, 2, 2, N, 3, N, 3]
   ex-2_1
Output: False
Explanation: Tree is not mirror image of itself i.e. tree is not symmetric.
Constraints:
1<= number of nodes<=2000